requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Flask==2.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.31
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.41
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
